related to #52 added a test
that uploads 3 revisions, waits and then continues the next revisions (with the goal of sending stubs of the attachment the second time, not the entire attachments.) Given what is in the code now it seems like it might be sending the entire attachment everytime.

I cant seem to get it to wait or call two replicators, i always get three revisions on the second replication.