for #52 convincing myself that we have to handroll
It seems we cant use standard libraries org.apache.http.entity.mime.MultipartEntity for multipart upload until CouchDB supports Content-disposition (minimally). There are lots of notes and logs in this commit which i used to try to trackdown where the multipart was failing (if on couchdb or in the testing code). My conclusion is that yeah, there is a high chance we can't use the standard multipart upload libraries.

Workarounds:
* hand roll the multipart with a byte stream
* do some sort of workaround to upload one attachment at a time.
* set up a bleading edge version of couchdb and see if it really can support standard multipart, then give up on workarounds and say everyone who wants to use this in their TouchDB has to have CouchDB version x...

The first two directions require significant changes to TouchDB. Unlike the iOS version which it looks like has streams, the Android code is passing around byte arrays.
Related commits in the iOS version:
https://github.com/couchbaselabs/TouchDB-iOS/commit/2bcc5167fdf4cad50ae285da393a427fcd387d53